Improve documentation for Spring Boot Applications
In case of Spring Boot application, the code example of the fixture is slightly different. The difference is so small that it is hard to catch. I tried to improve the Readme by adding a comment in the fixture code example to document that difference and call developer eyes to that place.